+++ tests/utils/asserts.ts
@@ -24,7 +24,7 @@ export const defaultStore: SerializedStore = {
       'sys:flavour': 'group',
       'sys:id': '1',
       'sys:children': ['2'],
-      'prop:xywh': '[50,50,300,200]',
+      'prop:xywh': '[50,50,720,480]',
     },
     '2': {
       'sys:flavour': 'paragraph',