+++ packages/editor/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @blocksuite/editor
 
+## 0.2.14
+
+### Patch Changes
+
+- b53f050: - feat: support edgeless mode click state
+  - fest: markdown import improvement
+  - fix: icon size
+  - fix: native text range selection improvement
+  - fix: quote format import support
+  - fix: page title enter in middle
+- Updated dependencies [b53f050]
+  - @blocksuite/blocks@0.2.14
+  - @blocksuite/store@0.2.14
+
 ## 0.2.11
 
 ### Patch Changes